+++ src/components/Toast/const.ts
@@ -1,0 +1,1 @@
+export const DEFAULT_TOAST_DURATION = 3000
